[refactor] 코드 분리
