+++ app/src/main/java/com/kt/startkit/ui/features/onboarding/OnBoardingScreen.kt
@@ -1,12 +1,12 @@
 package com.kt.startkit.ui.features.onboarding
 
 import androidx.compose.runtime.Composable
-import com.kt.startkit.ui.features.onboarding.common.Stepper
-import com.kt.startkit.ui.features.onboarding.find_pokemon.step.FindPokemonStep
+import com.kt.startkit.ui.features.onboarding.step.find_pokemon.sub_step.FindPokemonStep
+import com.kt.startkit.ui.features.onboarding.step.find_pokemon.sub_step.MajorStep
 
 @Composable
 fun OnBoardingScreen() {
-    Stepper(
-        steps = listOf(FindPokemonStep()),
+    OnBoardingStepper(
+        steps = MajorStep.values().map { it.step }
     )
 }
