+++ app/src/main/java/com/kt/startkit/ui/features/onboarding/step/find_pokemon/view/IntroView.kt
@@ -1,8 +1,8 @@
-package com.kt.startkit.ui.features.onboarding.find_pokemon.view
+package com.kt.startkit.ui.features.onboarding.step.find_pokemon.view
 
 import androidx.compose.runtime.Composable
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.kt.startkit.ui.features.onboarding.find_pokemon.FindPokemonViewModel
+import com.kt.startkit.ui.features.onboarding.step.find_pokemon.FindPokemonViewModel
 
 @Composable
 fun IntroView(
