+++ app/src/main/java/com/kt/startkit/ui/features/onboarding/step/find_pokemon/view/TravelSiteView.kt
@@ -1,13 +1,13 @@
-package com.kt.startkit.ui.features.onboarding.find_pokemon.view
+package com.kt.startkit.ui.features.onboarding.step.find_pokemon.view
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.hilt.navigation.compose.hiltViewModel
-import com.kt.startkit.ui.features.onboarding.find_pokemon.FindPokemonViewModel
+import com.kt.startkit.ui.features.onboarding.step.find_pokemon.FindPokemonViewModel
 
 @Composable
-fun RoutineView(
+fun TravelSiteView(
     findPokemonViewModel: FindPokemonViewModel,
 ) {
-    Text("routine")
+    Text("travel site")
 }
